# w-dms/swagger-ui/swagger.json
{"swagger":"2.0","info":{"description":"Provides API to operate with tasks","title":"Task API."},"paths":{"/task":{"get":{"description":"Reads tasks from database","produces":["application/json"],"operationId":"readTasks","parameters":[{"type":"string","x-go-name":"ID","description":"Id of the task to be found (skip to find all)","name":"id","in":"query"},{"type":"string","x-go-name":"Alias","description":"Alias of the task to be found (skip to find all)","name":"alias","in":"query"}],"responses":{"200":{"$ref":"#/responses/taskResponse"},"404":{"$ref":"#/responses/taskNotFound"}}},"put":{"description":"Updates task in database","consumes":["application/json"],"operationId":"updateTask","parameters":[{"name":"Body","in":"body","schema":{"$ref":"#/definitions/swaggerTask"}}],"responses":{"200":{"$ref":"#/responses/taskUpdated"},"404":{"$ref":"#/responses/taskNotFound"}}},"post":{"description":"Saves a task in the database","consumes":["application/json"],"produces":["No content"],"operationId":"saveTask","parameters":[{"name":"Body","in":"body","schema":{"$ref":"#/definitions/swaggerTask"}}],"responses":{"201":{"$ref":"#/responses/taskCreated"}}},"delete":{"description":"Deletes task from database","operationId":"deleteTask","parameters":[{"type":"string","x-go-name":"ID","description":"Id of the task to be found (skip to find all)","name":"id","in":"query"},{"type":"string","x-go-name":"Alias","description":"Alias of the task to be found (skip to find all)","name":"alias","in":"query"}],"responses":{"204":{"$ref":"#/responses/taskDeleted"},"404":{"$ref":"#/responses/taskNotFound"}}}}},"definitions":{"swaggerTask":{"type":"object","required":["alias","type"],"properties":{"alias":{"description":"Task Alias","type":"string","x-go-name":"Alias"},"desc":{"description":"Task description","type":"string","x-go-name":"Description"},"etime":{"description":"Task estimation time","type":"string","x-go-name":"EstimateTime"},"id":{"description":"Task Id","type":"string","x-go-name":"ID"},"reminders":{"description":"Times when the task should be reminded","type":"array","items":{"type":"string"},"x-go-name":"Reminders"},"rtime":{"description":"Task actual time","type":"string","x-go-name":"RealTime"},"tags":{"description":"Task tags","type":"array","items":{"type":"string"},"x-go-name":"Tags"},"ts":{"description":"Timastamp when the task was created","type":"integer","format":"int64","x-go-name":"Timestamp"},"type":{"description":"Task type","type":"string","x-go-name":"Type"}},"x-go-package":"github.com/arithmetics/w-dms"}},"responses":{"taskCreated":{"description":"Task created"},"taskDeleted":{"description":"Task deleted"},"taskNotFound":{"description":"Task was not found"},"taskResponse":{"description":"Tasks were read successfully","schema":{"type":"array","items":{"$ref":"#/definitions/swaggerTask"}}},"taskUpdated":{"description":"Task updated"}}}